# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Kimthean_go-chat_internal_types.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_Kimthean_go-chat_internal_types.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_Kimthean_go-chat_internal_types.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: Langchain go RAG agent API.
  termsOfService: http://swagger.io/terms/
  title: LangChain Go RAG Agent API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Log in a user with email and password
      parameters:
      - description: Login user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_Kimthean_go-chat_internal_types.LoginRequest'
      produces:
      - application/json
      responses: {}
      summary: Log in a user
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Log out a user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/github_com_Kimthean_go-chat_internal_types.RefreshTokenRequest'
      produces:
      - application/json
      responses: {}
      summary: Refresh access token
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign up user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_Kimthean_go-chat_internal_types.SignUpRequest'
      produces:
      - application/json
      responses: {}
      summary: Sign up a new user
      tags:
      - auth
  /user/me:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: get user details
      tags:
      - user
  /user/profile:
    post:
      parameters:
      - description: Profile Image
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: upload profile image
      tags:
      - user
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"